+++ R/HALT.R
@@ -12,27 +12,27 @@
 #' consider using \code{\link{HALT_standalone}()}.
 #' @param label (Character scalar) Label to give the HALT results in the output file.
 #' @param config (object of class HALT_config or path to config file generated with HALTConfig)
-#' Use the functions auto_config() or make_config() to generate this or provide a filename of a config file generated with the HALTConfig Shiny App
+#' Use the functions \code{\link[=auto_config]{auto_config()}} or \code{\link[=make_config]{make_config()}} to generate this or provide a filename of a config file generated with the HALTConfig Shiny App
 #' See also the documentation there for further explanations.
 #' @param audio_dir (url). The URL of the directory containing the stimuli.
-#' @param no_screening (Scalar boolean). Whether to include screening parts or not.
+#' @param show_id (boolean default = F) Flag whether question IDs should be displayed (e.g. for debugging)
 #' @param dict The psychTestR dictionary used for internationalisation.
 #' @export
 HALT <- function(label = "HALT",
                  config = HALT::auto_config(),
                  audio_dir = "https://media.gold-msi.org/test_materials/HLT",
-                 no_screening = FALSE,
+                 show_id = FALSE,
                  dict = HALT::HALT_dict) {
   stopifnot(is(config, "HALT_config") || is.character(config) && file.exists(config))
   stopifnot(purrr::is_scalar_character(label))
   stopifnot(purrr::is_scalar_character(audio_dir))
 
-
-  audio_dir <- gsub("/$", "", audio_dir)
-  main_test <-  main_test(label = label, audio_dir = audio_dir, config, dict = dict, type = config$volume_level)
-  if(no_screening){
-    main_test <- main_test_no_screening(label = label, audio_dir = audio_dir, config, dict = dict, type = config$volume_level)
+  if(is.character(config) && file.exists(config)) {
+    config <- import_config(config)
   }
+  audio_dir <- gsub("/$", "", audio_dir)
+  type <- c("-8.4 LUFS" = "loud",  "-20.0 LUFS" = "quiet")[config$volume_level] %>% as.vector()
+  main_test <-  main_test(label = label, audio_dir = audio_dir, config, dict = dict, type = type, show_id = show_id)
 
   psychTestR::join(
     psychTestR::begin_module(label),