+++ R/item_page.R
@@ -52,6 +52,7 @@ get_audio_ui <- function(url,
     oncanplaythrough = media_js$show_media_btn,
     onplay = paste0(media_js$media_played, media_js$hide_media_btn),
     controls = if (show_controls) "controls",
+    controlsList = "nodownload",
     #onended = if (wait) paste0(media_js$show_responses, media_js$hide_media) else "null",
     onended = if (wait) media_js$show_responses else "null"
   )
@@ -139,6 +140,7 @@ audio_text_page <- function(page_no,
                             correct_answer = "",
                             save_answer = TRUE,
                             on_complete = NULL,
+                            show_id = FALSE,
                             admin_ui = NULL) {
 
   stopifnot(purrr::is_scalar_integer(page_no))
@@ -167,7 +169,7 @@ audio_text_page <- function(page_no,
   }
   response_ui <- shiny::div(
     shiny::p(
-      shiny::textInput("text_input", label = label, placeholder = "", width = "50"),
+      shiny::textInput("text_input", label = ifelse(show_id, label, ""), placeholder = "", width = "50"),
       psychTestR::trigger_button("next", psychTestR::i18n("CONTINUE"))),
     id = "response_ui", style = "visibility:hidden")
 
@@ -218,16 +220,18 @@ HALT_audio_NAFC_page <- function(page_no,
                                  audio_dir,
                                  save_answer = TRUE,
                                  admin_ui = NULL,
-                                 type = "loud"){
+                                 type = "loud",
+                                 config,
+                                 show_id = FALSE){
   label <- sprintf("po%d%s", page_no, sub_id)
   stopifnot(purrr::is_scalar_character(label))
   audio_url <- get_audio_url(audio_dir, page_no, sub_id, type = type)
   correct_answers <- get_item(page_no, sub_id, "correct_answer", type = type)
   messagef("Page no: %d, sub id = %s, correct: %s", page_no, sub_id, correct_answers)
-  prompt <- shiny::div(get_page_counter(page_no, num_pages, ABC_offset),
+  prompt <- shiny::div(get_page_counter(page_no, num_pages, ABC_offset, config),
                        shiny::div(
                          psychTestR::i18n(sprintf("THLT_%04d_PROMPT", page_no)),
-                         label,
+                         ifelse(show_id, label, ""),
                          style  = HALT_standard_style),
                        shiny::p(""))
   on_complete <- function(answer, state, ...) {
@@ -282,19 +286,34 @@ select_left_right <- function(direction){
 }
 
 
-get_page_counter <- function(page_no, num_pages, ABC_offset = 0L, no_screening = FALSE){
+get_page_counter <- function(page_no, num_pages, ABC_offset = 0L, config){
   orig_page <- page_no
   if(is.null(num_pages) || is.na(num_pages) || num_pages < 1){
     return(shiny::div(""))
   }
-  counter_map <-c("1" = 1L, "2" = 2L, "3" = 3L, "4" = 4L, "5" = 5L, "device" = 6L, "6" = 7L,
-                  "7" = 13L, "13" = 19L, "8" = 26L, "9" = 27L, "10" = 28L, "11" = 29L)
-  if(no_screening){
-    counter_map <-c("1" = 1L, "2" = 2L, "3" = 3L, "4" = 4L, "5" = 5L, "device" = 6L,
-                    "8" = 7L, "9" = 8L, "10" = 9L, "11" = 10L)
-
+  counter_map <-c("1" = 1L, "2" = 2L)
+  if (config$channel_check) {
+    counter_map <- append(counter_map, c("4" = 3L, "5" = 4L, "device" = 5L))
+    if (config$screening_parts) {
+      counter_map <- append(counter_map, c("6" = 6L, "7" = 12L, "13" = 18L))
+      if(config$frequency_check) {
+        counter_map <- append(counter_map,
+                              c("8" = 24L, "9" = 25L, "10" = 26L, "11" = 27L))
+      }
+    } else {
+      if (config$frequency_check) {
+        counter_map <- append(counter_map,
+                              c("8" = 6L, "9" = 7L, "10" = 8L, "11" = 9L))
+      }
+    }
+  } else {
+    counter_map <- append(counter_map, c("device" = 3L))
+    if (config$frequency_check) {
+      counter_map <- append(counter_map,
+                            c("8" = 4L, "9" = 5L, "10" = 6L, "11" = 7L))
+    }
   }
-  num_pages <- max(counter_map)
+  #num_pages <- max(counter_map)
   counter <- as.character(counter_map[as.character(page_no)] + ABC_offset)
   #messagef("Page_no %s, num_pages = %d, ABC_offset = %d, counter = %s", page_no, num_pages, ABC_offset, counter)
 
@@ -311,14 +330,15 @@ warning_page <- function(label, warning_message){
     dict = HALT::HALT_dict)
 }
 
-HALT_base_page <- function(page_no, sub_id, num_pages, audio_dir, save_answer = T, no_screening = FALSE, type = "loud"){
+HALT_base_page <- function(page_no, sub_id, num_pages, audio_dir, save_answer = T, config = HALT::auto_config(), type = "loud", show_id = FALSE){
   audio_text_page(page_no, sub_id,
-                  prompt = shiny::div(get_page_counter(page_no, num_pages, no_screening = no_screening),
+                  prompt = shiny::div(get_page_counter(page_no, num_pages, config = config),
                                       shiny::div(psychTestR::i18n(sprintf("THLT_%04d_PROMPT", page_no)),
                                                  style = HALT_standard_style),
                                       shiny::p("")),
                   audio_url = get_audio_url(audio_dir, page_no, sub_id, type = type),
                   correct_answer = get_item(page_no, sub_id, "correct_answer", type = type),
+                  show_id = show_id,
                   save_answer = save_answer)
 }
 
@@ -374,7 +394,7 @@ test_force_loop <- function(page_no, max_count, correct_answer = "correct"){
 }
 
 
-left_right_page <- function(audio_dir, right_first, num_pages, no_screening = FALSE){
+left_right_page <- function(audio_dir, right_first, num_pages, config = HALT::auto_config()){
   perm <- c(1L, 2L)
   if(right_first){
     perm <- c(2L, 1L)
@@ -391,7 +411,7 @@ left_right_page <- function(audio_dir, right_first, num_pages, no_screening = FA
   }
   psychTestR::new_timeline(
     psychTestR::NAFC_page("po4",
-                          prompt = shiny::div(get_page_counter(4L, num_pages, no_screening = no_screening),
+                          prompt = shiny::div(get_page_counter(4L, num_pages, config = config),
                                               shiny::div(psychTestR::i18n("THLT_0004_PROMPT"),
                                                          style = HALT_standard_style),
                                               shiny::p(shiny::img(src = img_url,
@@ -431,7 +451,7 @@ device_page <- function(num_pages, config){
 
   psychTestR::new_timeline(
     psychTestR::NAFC_page("device_screening",
-                          prompt = shiny::div(get_page_counter("device", num_pages),
+                          prompt = shiny::div(get_page_counter("device", num_pages, config = config),
                                               shiny::div(psychTestR::i18n("DEVICE_PROMPT"),
                                                          style = HALT_standard_style)),
                           choices = as.character(1:6),
@@ -454,13 +474,13 @@ scc_page <- function(dict = HALT::HALT_dict, config){
 }
 
 #Creating main pages functions
-page_po1 <- function(audio_dir, num_pages, no_screening = FALSE, type = "loud"){
+page_po1 <- function(audio_dir, num_pages, config = HALT::auto_config(), type = "loud"){
   messagef("page_po1: %d", num_pages)
 
   psychTestR::new_timeline(
     volume_calibration_page(
       url = get_audio_url(audio_dir, 1L, "", type = type),
-      prompt = shiny::div(get_page_counter(1L, num_pages, no_screening = no_screening),
+      prompt = shiny::div(get_page_counter(1L, num_pages, config = config),
                           shiny::div(psychTestR::i18n("THLT_0001_PROMPT"),
                                      style = HALT_standard_style),
                           shiny::p("")),
@@ -497,26 +517,26 @@ page_po4_old <- function(audio_dir, config, num_pages){
       )))
 }
 
-page_po4 <- function(config, audio_dir, num_pages, no_screening = FALSE){
+page_po4 <- function(config, audio_dir, num_pages){
   messagef("page_po4: %d", num_pages)
   psychTestR::join(
       psychTestR::conditional(test = select_left_right("left"),
-                              logic = left_right_page(audio_dir, right_first = F, num_pages, no_screening = no_screening)),
+                              logic = left_right_page(audio_dir, right_first = F, num_pages, config = config)),
       psychTestR::conditional(test = select_left_right("right"),
-                              logic = left_right_page(audio_dir, right_first = T, num_pages, no_screening = no_screening)),
+                              logic = left_right_page(audio_dir, right_first = T, num_pages, config = config)),
       psychTestR::conditional(test = test_answer(page_no = 4L, config, "right"),
                               logic = HALT_stop_page()))
 }
 
 
-page_po5 <- function(config, audio_dir, num_pages, no_screening = FALSE, type = "loud"){
+page_po5 <- function(config, audio_dir, num_pages, type = "loud", show_id = FALSE){
   psychTestR::join(
-    page_calibrate(page_no = 5L, num_pages, audio_dir = audio_dir, save_answer = T, no_screening = no_screening, type = type),
+    page_calibrate(page_no = 5L, num_pages, audio_dir = audio_dir, save_answer = T, config = config, type = type, show_id = show_id),
     psychTestR::conditional(test = test_answer(page_no = 5L, config, "stereo channels correct", invert = T),
                             logic = HALT_stop_page()))
 }
 
-page_force_correct <- function(page_no, num_pages, config, audio_dir, no_screening = FALSE, type = "loud"){
+page_force_correct <- function(page_no, num_pages, config, audio_dir, type = "loud", show_id = FALSE){
   warning_label <- sprintf("warning_po%d", page_no)
   psychTestR::join(
     psychTestR::code_block(
@@ -532,7 +552,7 @@ page_force_correct <- function(page_no, num_pages, config, audio_dir, no_screeni
                                 logic = warning_page(warning_label, "WARNING_TOO_QUIET")),
         psychTestR::conditional(test = test_answer(page_no, config, "imprecise"),
                                 logic = warning_page(warning_label, "WARNING_IMPRECISE")),
-        page_calibrate(page_no = page_no, num_pages, audio_dir = audio_dir, save_answer = T, no_screening = no_screening, type = type),
+        page_calibrate(page_no = page_no, num_pages, audio_dir = audio_dir, save_answer = T, config = config, type = type, show_id = show_id),
         psychTestR::code_block(function(state, ...){
           counter <- psychTestR::get_local(key = sprintf("po%d_counter", page_no), state)
           psychTestR::set_local(key = sprintf("po%d_counter", page_no), value = counter + 1L, state)
@@ -541,7 +561,7 @@ page_force_correct <- function(page_no, num_pages, config, audio_dir, no_screeni
     ))
 }
 
-page_calibrate <- function(page_no, num_pages, audio_dir, save_answer = T, no_screening = FALSE, type = "loud"){
+page_calibrate <- function(page_no, num_pages, audio_dir, save_answer = T, config = HALT::auto_config(), type = "loud", show_id = F){
   psychTestR::new_timeline(
     psychTestR::join(
       psychTestR::code_block(function(state, ...){
@@ -552,12 +572,12 @@ page_calibrate <- function(page_no, num_pages, audio_dir, save_answer = T, no_sc
       }),
       psychTestR::reactive_page(function(state, ...){
         selection <- psychTestR::get_local("current_selection", state)
-        HALT_base_page(page_no, selection, num_pages, audio_dir, save_answer, no_screening = no_screening, type = type)
+        HALT_base_page(page_no, selection, num_pages, audio_dir, save_answer, config = config, type = type, show_id = show_id)
       }))
     , dict = HALT::HALT_dict)
 }
 
-page_ABC_section <- function(page_no, num_pages, audio_dir, type = "loud"){
+page_ABC_section <- function(page_no, num_pages, audio_dir, type = "loud", config, show_id = FALSE){
   psychTestR::new_timeline(
     psychTestR::join(
       psychTestR::code_block(
@@ -584,7 +604,9 @@ page_ABC_section <- function(page_no, num_pages, audio_dir, type = "loud"){
                                  num_pages, counter,
                                  audio_dir = audio_dir,
                                  save_answer = T,
-                                 type = type)
+                                 type = type,
+                                 config = config,
+                                 show_id = show_id)
           }
           ),
           psychTestR::code_block(function(state, ...) {