+++ R/prevalence_calc.R
@@ -1,7 +1,7 @@
-#' Maximum Headphones Predictive Value
+#' Calculate predictive values and overall utility
 #'
-#' This function returns a data.frame containing all tests with maximum
-#' headphones predictive value for the specified prevalence of headphones.
+#' This function calculates the predictive values and the overall utility for a
+#' given prevalence for headphones.
 #'
 #' @param baserate_hp Sets the (estimated) prevalence of headphones in the
 #' target population as a number between 0 and 1. Defaults to the unbiased
@@ -11,7 +11,7 @@
 #' \insertRef{HALTpaper}{HALT}
 #'
 #' @export
-max_hp_pv <- function(baserate_hp = 211/1194) {
+tests_pv_utility <- function(baserate_hp = 211/1194) {
   stopifnot(baserate_hp < 1,
             baserate_hp > 0)
   tests <- HALT::test_config
@@ -21,8 +21,24 @@ max_hp_pv <- function(baserate_hp = 211/1194) {
     (1 - baserate_hp) * tests$true_ls_rate / ((1 - baserate_hp) * tests$true_ls_rate + baserate_hp * tests$false_ls_rate)
   tests$utility <-
     baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$true_ls_rate
-
-  tests <- tests %>% filter(hp_pv == max(hp_pv)) %>% dplyr::select(method, method_code, A, B, C, true_hp_rate, true_ls_rate, hp_pv, ls_pv, utility)
+#  tests <- tests %>% dplyr::select(method, method_code, A, B, C, true_hp_rate, true_ls_rate, hp_pv, ls_pv, utility)
+  tests
+}
+#' Maximum Headphones Predictive Value
+#'
+#' This function returns a data.frame containing all tests with maximum
+#' headphones predictive value for the specified prevalence of headphones.
+#'
+#' @inheritParams tests_pv_utility
+#'
+#' @references
+#' \insertRef{HALTpaper}{HALT}
+#'
+#' @export
+max_hp_pv <- function(baserate_hp = 211/1194) {
+  stopifnot(baserate_hp < 1,
+            baserate_hp > 0)
+  tests <- tests_pv_utility(baserate_hp = baserate_hp) %>% filter(hp_pv == max(hp_pv))
   tests
 }
 #' Maximum Loudspeakers Predictive Value
@@ -30,7 +46,7 @@ max_hp_pv <- function(baserate_hp = 211/1194) {
 #' This function returns a data.frame containing all tests with maximum
 #' loudspeakers predictive value for the specified prevalence of headphones.
 #'
-#' @inheritParams max_hp_pv
+#' @inheritParams tests_pv_utility
 #'
 #' @references
 #' \insertRef{HALTpaper}{HALT}
@@ -39,15 +55,7 @@ max_hp_pv <- function(baserate_hp = 211/1194) {
 max_ls_pv <- function(baserate_hp = 211/1194) {
   stopifnot(baserate_hp < 1,
             baserate_hp > 0)
-  tests <- HALT::test_config
-  tests$hp_pv <-
-    baserate_hp * tests$true_hp_rate / (baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$false_hp_rate)
-  tests$ls_pv <-
-    (1 - baserate_hp) * tests$true_ls_rate / ((1 - baserate_hp) * tests$true_ls_rate + baserate_hp * tests$false_ls_rate)
-  tests$utility <-
-    baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$true_ls_rate
-
-  tests <- tests %>% filter(ls_pv == max(ls_pv)) %>% dplyr::select(method, method_code, A, B, C, true_hp_rate, true_ls_rate, hp_pv, ls_pv, utility)
+  tests <- tests_pv_utility(baserate_hp = baserate_hp) %>% filter(ls_pv == max(ls_pv))
   tests
 }
 #' Maximum Utility maximizing percent correct
@@ -56,7 +64,7 @@ max_ls_pv <- function(baserate_hp = 211/1194) {
 #' for utility maximizing percent correct for the specified prevalence of
 #' headphones.
 #'
-#' @inheritParams max_hp_pv
+#' @inheritParams tests_pv_utility
 #'
 #' @references
 #' \insertRef{HALTpaper}{HALT}
@@ -65,15 +73,7 @@ max_ls_pv <- function(baserate_hp = 211/1194) {
 max_utility <- function(baserate_hp = 211/1194) {
   stopifnot(baserate_hp < 1,
             baserate_hp > 0)
-  tests <- HALT::test_config
-  tests$hp_pv <-
-    baserate_hp * tests$true_hp_rate / (baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$false_hp_rate)
-  tests$ls_pv <-
-    (1 - baserate_hp) * tests$true_ls_rate / ((1 - baserate_hp) * tests$true_ls_rate + baserate_hp * tests$false_ls_rate)
-  tests$utility <-
-    baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$true_ls_rate
-
-  tests <- tests %>% filter(utility == max(utility)) %>% dplyr::select(method, method_code, A, B, C, true_hp_rate, true_ls_rate, hp_pv, ls_pv, utility)
+  tests <- tests_pv_utility(baserate_hp = baserate_hp) %>% filter(utility == max(utility))
   tests
 }
 #' Test procedures with maximum properties
@@ -83,7 +83,7 @@ max_utility <- function(baserate_hp = 211/1194) {
 #' or maximum value for utility maximizing percent correct for the specified
 #' prevalence of headphones.
 #'
-#' @inheritParams max_hp_pv
+#' @inheritParams tests_pv_utility
 #'
 #' @references
 #' \insertRef{HALTpaper}{HALT}
@@ -220,13 +220,7 @@ a_priori_est <- function(baserate_hp = 211/1194,
             tolerance >= 0,
             as.integer(tolerance)==as.double(tolerance))
 
-  tests <- HALT::test_config
-  tests$hp_pv <-
-    baserate_hp * tests$true_hp_rate / (baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$false_hp_rate)
-  tests$ls_pv <-
-    (1 - baserate_hp) * tests$true_ls_rate / ((1 - baserate_hp) * tests$true_ls_rate + baserate_hp * tests$false_ls_rate)
-  tests$utility <-
-    baserate_hp * tests$true_hp_rate + (1 - baserate_hp) * tests$true_ls_rate
+  tests <- tests_pv_utility(baserate_hp = baserate_hp)
 
   q <- (qnorm(p = 1 - min_prob))^2